Moved Controller, Entity and Form folders to the proper place
