+++ foo_bar.routing.yml
@@ -36,7 +36,7 @@ foo_bar.delete:
 foo_bar.settings:
   path: 'admin/config/foo_bar_settings'
   defaults:
-   _form: '\Drupal\foo_bar\Entity\Form\FooBarSettingsForm'
+   _form: '\Drupal\foo_bar\Form\FooBarSettingsForm'
    _title: 'FooBar Settings'
   requirements:
     _permission: 'admin_foo_bar'
